Initial work on extendable images (#9)
We now implement a custom loader which lets us load templates as:
  - '<imageid>/pxedata'
  - '<imageid>/script'
  - '<imageid>/postinstall'
  - etc.

This means we can now use twig extends/includes in templates.

At the moment we still need to manually ensure that all the required variables
are added to the child image or else bits will be left blank.

In future we will automatically (somehow) figure out what variables
are needed by chaining upwards.