Add simple Dockerfile.
This is untested in a production environment, but should run the web app
and behave reasonably. YMMV.

Required volumes:
  - `/var/lib/tftpboot` - This is where tftp data will go.
  - `/pxemanager/data`  - data persistence (eg server and image data)

A `config.local.php` file can be bind-mounted to `/pxemanager/config.local.php`
if required.

The `/var/lib/tftpboot` volume should be shared with a tftp server that
actually serves the data. This will need to be pre-seeded manually with the
required pxelinux.0, ldlinux.c32 and ipxe.lkrn files.